Deep dive
A further deep dive into the ATOM package using the 20newsgroups dataset. We fetch articles across 5 categories, train a Random Forest classifier on them and predict the values of the test set. We employ various visualization techniques to help us determine which words in the articles helped in the classification of the article. An accuracy of 93% is achieved.